Add new icons and banner.
The old icons were hard to make out at their typical size. The new icons are
clearer, and avoid the lock iconography in order to avoid creating confusion
with the lock icon in the URL bar.

Add a new icon type "inactive" that indicates that no rules applied on a given
site.

Disables the counter by default on Firefox (you can still turn
it on manually), and adds a new icon state "inactive."